[PVM] Improve storage items deletion in `EJECT` hostcall (#174)
Updated `AccountsSandboxMap::eject_account` method to explicitly take the key of the last lookups entry of the ejecting account as an argument and then remove that from the global state, instead of iterating on all storage items on the ejecting account sandbox. The previous approach was redundant since the hostcall already checks no storage items should exist in the ejecting account other than the last one lookups entry.